refactor(db): extract checkpoint artifact lifecycle
